refactor: Update StudentCard component styles for responsive design

+++ src/app/api/v1/alumni/route.ts
@@ -1,18 +1,30 @@
-import { Alumni } from "@/types/Alumni";
+import { AlumniDashboard } from "@/types/Alumni";
 import { NextRequest, NextResponse } from "next/server";
 import { connect } from "@/dfConfig/dbConfig";
 import AlumniModel from "@/models/alumni";
+import jwt from 'jsonwebtoken';
 
 connect();
 
 export async function GET(req: NextRequest) {
     try {
-        const alumniList = await AlumniModel.find();
-        const alumniData: Alumni[] = alumniList.map(data => ({
+        const alumniList = await AlumniModel.find({ isVerified: true });
+        const alumniData: AlumniDashboard[] = alumniList.map(data => ({
             name: data.name,
             batch: data.batch,
             department: data.department,
-            photo: data.image // Assuming 'photo' in response is mapped to 'image' in the model
+            photo: data.photo,
+            email: data.email,
+            phone: data.phone,
+            address: data.address,
+            company: data.company,
+            position: data.position,
+            linkedin: data.linkedin,
+            rating: data.rating,
+            review: data.review,
+            isVerified: data.isVerified,
+            password: data.password,
+            isTestimonial:data.isTestimonial
         }));
         return NextResponse.json(alumniData, { status: 200 });
     } catch (error: any) {
@@ -21,53 +33,36 @@ export async function GET(req: NextRequest) {
     }
 }
 
-export async function POST(req: NextRequest) {
-    try {
-        const {
-            name, batch, email, phone, address,
-            company, position, image, linkedin,
-            department, rating, review
-        } = await req.json();
-
-        const newAlumni = new AlumniModel({
-            name, batch, email, phone, address,
-            company, position, image, linkedin,
-            department, rating, review,
-            isVerified: false
-        });
-
-        await newAlumni.save();
-        return NextResponse.json({ message: 'Alumni details added successfully' }, { status: 200 });
-    } catch (error: any) {
-        console.log('Error occurred @ /api/v1/alumni/route.ts POST method:->', error.message);
-        return NextResponse.json({ message: error.message }, { status: 500 });
-    }
-}
 
 export async function PUT(req: NextRequest) {
     try {
-        const email = req.nextUrl.searchParams.get('email');
+        const userToken = req.cookies.get('token')?.value;
+        if (!userToken) {
+            return NextResponse.json({ message: 'Login first' }, { status: 404 });
+        }
+
+        const decoded = jwt.decode(userToken) as { email?: string };
+        const email = decoded?.email;
+
         if (!email) {
-            return NextResponse.json({ message: 'Please provide an email' }, { status: 400 });
+            return NextResponse.json({ message: 'Email not found in token' }, { status: 400 });
         }
 
         const {
             name, batch, phone, address,
-            company, position, image, linkedin,
+            company, position, photo, linkedin,
             department, rating, review
         } = await req.json();
 
         const updatedAlumni = await AlumniModel.findOneAndUpdate(
             { email },
-            { name, batch, phone, address, company, position, image, linkedin, department, rating, review },
+            { name, batch, phone, address, company, position, photo: photo, linkedin, department, rating, review, isVerified: false },
             { new: true }
         );
-
         if (!updatedAlumni) {
             return NextResponse.json({ message: 'Alumni details not found' }, { status: 404 });
         }
-
-        return NextResponse.json({ message: 'Alumni details updated successfully', alumni: updatedAlumni }, { status: 200 });
+        return NextResponse.json({ message: 'Alumni details update requested', alumni: updatedAlumni }, { status: 200 });
     } catch (error: any) {
         console.log('Error occurred @ /api/v1/alumni/route.ts PUT method:->', error.message);
         return NextResponse.json({ message: error.message }, { status: 500 });
@@ -76,10 +71,20 @@ export async function PUT(req: NextRequest) {
 
 export async function DELETE(req: NextRequest) {
     try {
+        const userToken = req.cookies.get('token')?.value;
+        if (!userToken) {
+            return NextResponse.json({ message: 'Login first' }, { status: 404 });
+        }
+        const decodedToken = jwt.decode(userToken) as { admin?: {id?:string,admin?:boolean}};
+        if (decodedToken?.admin?.admin !== true) {
+            return NextResponse.json({ message: 'Admin not found' }, { status: 404 });
+        }
+
         const email = req.nextUrl.searchParams.get('email');
         if (!email) {
             return NextResponse.json({ message: 'Please provide an email' }, { status: 400 });
         }
+
 
         const deletedAlumni = await AlumniModel.findOneAndDelete({ email });
         if (!deletedAlumni) {
